Support belated initialization to zero for counters
Counters may be configured to do a belated initialization to zero. The
first time such a counter is collected, it reports a zero, even if it
was initialized with a different value. Together with the zero it
reports a timestamp back in time, depending on the configuration.
The actual counter value will only be reported from the second time on
it is collected.